+++ hpobench/benchmarks/ml/rf_benchmark.py
@@ -6,7 +6,7 @@ import numpy as np
 from ConfigSpace.hyperparameters import Hyperparameter
 from sklearn.ensemble import RandomForestClassifier
 
-from hpobench.dependencies.ml_mmfb.ml_benchmark_template import MLBenchmark
+from hpobench.dependencies.ml.ml_benchmark_template import MLBenchmark
 
 
 class RandomForestBenchmark(MLBenchmark):
